+++ src/components/OrganizationSaas/Dashboard/index.tsx
@@ -1,0 +1,11 @@
+import { useOrganization } from '@vocdoni/react-providers'
+
+const OrganizationDashboard = () => {
+  const { organization } = useOrganization()
+
+  if (!organization) return null
+
+  return null
+}
+
+export default OrganizationDashboard
